# docs/swagger.yaml
swagger: "2.0"
info:
 
  version: "1.0.0"
  title: "Story Management"
  
host: "localhost:5000/api"
basePath: "/sm"
tags:
- name: "Story Management"
  
schemes:
- "http"
paths:
 
  /createStory:
    post:
      tags:
      - "Story Management"
      summary: "Create new Story"
      description: "Create new Story for the current user"
      operationId: "CreateStory"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Object to create Story"
        required: true
        schema:
          $ref: "#/definitions/CreateStory"
      - in: "header"
        name: "auth-token"
        description: "AuthToken Generated From Login EndPoint"
        required: true
        schema:
          $ref: "#/definitions/AuthToken"
      responses:
        201:
          description: "New Story Created Successfully"
        400:
          description: "Invalid Request Data"
        401:
          description: "Unauthorized"
  /stories:
    get:
      tags:
      - "Story Management"
      summary: "Get All Stories"
      description: "List of stories from all users"
      operationId: "GetAllStory"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "auth-token"
        description: "AuthToken Generated From Login EndPoint"
        required: true
        schema:
          $ref: "#/definitions/AuthToken"
      responses:
        200:
          description: "Get List Of Stories"
        400:
          description: "Invalid Request Data"
        401:
          description: "Unauthorized"
  
  /myStories:
    get:
      tags:
      - "Story Management"
      summary: "Get Stories of current user"
      description: "List of stories associated with currently logged in user"
      operationId: "MyStories"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "auth-token"
        description: "AuthToken Generated From Login EndPoint"
        required: true
        schema:
          $ref: "#/definitions/AuthToken"
      responses:
        200:
          description: "List of stories"
        400:
          description: "Invalid Request Data"
        401:
          description: "Unauthorized"
          
  /deleteStory:
    post:
      tags:
      - "Story Management"
      summary: "Delete a Story of current user"
      operationId: "DeleteStory"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Object with userId and storyId"
        required: true
        schema:
          $ref: "#/definitions/DeleteStory"
      - in: "header"
        name: "auth-token"
        description: "AuthToken Generated From Login EndPoint"
        required: true
        schema:
          $ref: "#/definitions/AuthToken"
      responses:
        200:
          description: "List of stories"
        400:
          description: "Invalid Request Data"
        401:
          description: "Unauthorized"
  
definitions:
  AuthToken:
      type: "object"
      properties:
        auht-token:
          type: "string"
      xml:
        name: "Auth"
  CreateStory:
    type: "object"
    properties:
      title:
        type: "string"
      storyBody:
        type: "string"
    xml:
      name: "CreateStory"
  DeleteStory:
    type: "object"
    properties:
      storyId:
        type: "string"
    xml:
      name: "MyStories"
  